requirements.txt
# --- STT (Speech-to-Text) ---
faster-whisper==1.0.3
numpy>=1.23.0
scipy>=1.10.0
sounddevice>=0.4.6

# --- Robot control ---
pyserial>=3.5
pydobot==1.3.3   # ใช้เวอร์ชันที่สื่อสารกับ Dobot ได้ (เปลี่ยนตามที่ conda env ของคุณใช้ได้จริง)

# --- Utilities ---
setuptools<81    # ป้องกัน warning pkg_resources